fix: fix timestamp bugs

+++ src/driver_hcsr04.c
@@ -286,8 +286,8 @@ uint8_t hcsr04_read(hcsr04_handle_t *handle, uint32_t *time_us, float *m)
            
             return 1;                                                                 /* return error */
         }
-        *time_us = (time_stop.millisecond - time_start.millisecond) * 1000
-                   +(int32_t)(time_stop.microsecond - time_start.microsecond);        /* get time */
+        *time_us = (uint32_t)((time_stop.millisecond - time_start.millisecond) * 1000
+                   +(int32_t)(time_stop.microsecond - time_start.microsecond));       /* get time */
         if ((*time_us) > 150 * 1000)                                                  /* check time */
         {
             if (retry != 0)                                                           /* check remain retry times */
